Fix sforl
Reorganise files.
