+++ internal/sforl/doc.go
@@ -1,0 +1,3 @@
+// Package sforl contains abstractions of the Set for Life draw from UK National lottery and
+// all related data analytics.
+package sforl
