--- internal/sforl/sqlite.go
@@ -1,85 +1,0 @@
-package sforl
-
-import (
-	"context"
-	"database/sql"
-	"fmt"
-	"log"
-	"paulwizviz/lotterystat/internal/dbutil"
-	"time"
-)
-
-var (
-	createSQLiteTableSQL   = fmt.Sprintf(`CREATE TABLE IF NOT EXISTS %s (%s INTEGER,%s INTEGER,%s INTEGER,%s INTEGER,%s INTEGER,%s INTEGER,%s INTEGER,%s INTEGER, %s TEXT,%s TEXT,%s INTEGER PRIMARY KEY)`, tblName, drawDate, dayOfWeek, ball1, ball2, ball3, ball4, ball5, luckyBall, ballset, machine, drawNo)
-	insertSQLiteDrawSQL    = fmt.Sprintf(`INSERT INTO %s (%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s) VALUES ( ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ? )`, tblName, drawDate, dayOfWeek, ball1, ball2, ball3, ball4, ball5, luckyBall, ballset, machine, drawNo)
-	selectSQLiteAllDrawSQL = fmt.Sprintf(`SELECT * FROM %s`, tblName)
-)
-
-func CreateSQLiteTable(ctx context.Context, db *sql.DB) error {
-	return createSQLiteTable(ctx, db)
-}
-
-func createSQLiteTable(ctx context.Context, db *sql.DB) error {
-	_, err := db.ExecContext(ctx, createSQLiteTableSQL)
-	if err != nil {
-		return fmt.Errorf("%w-%s", dbutil.ErrDBCreateTbl, err.Error())
-	}
-	return nil
-}
-
-func PersistsCSV(ctx context.Context, db *sql.DB, nworkers int) error {
-	return persistsCSV(ctx, db, nworkers)
-}
-
-func listSQLiteAllDraw(ctx context.Context, db *sql.DB) ([]Draw, error) {
-	var draws []Draw
-	row, err := db.QueryContext(ctx, selectSQLiteAllDrawSQL)
-	if err != nil {
-		return nil, fmt.Errorf("%w-%s", dbutil.ErrDBQueryTbl, err.Error())
-	}
-	for row.Next() {
-		d := Draw{}
-		var drawDate int
-		err := row.Scan(&drawDate, &d.DayOfWeek, &d.Ball1, &d.Ball2, &d.Ball3, &d.Ball4, &d.Ball5, &d.LifeBall, &d.BallSet, &d.Machine, &d.DrawNo)
-		if err != nil {
-			return nil, fmt.Errorf("%w-%s", dbutil.ErrDBQueryTbl, err.Error())
-		}
-		d.DrawDate = time.Unix(int64(drawDate), 0)
-		draws = append(draws, d)
-	}
-	return draws, nil
-}
-
-func persistsSQLiteDraw(ctx context.Context, db *sql.DB, dc <-chan DrawChan) error {
-	stmt, err := prepInsertSQLiteDrawStmt(ctx, db)
-	if err != nil {
-		return err
-	}
-	for c := range dc {
-		if c.Err != nil {
-			continue
-		}
-		_, err = insertSQLiteDraw(ctx, stmt, c.Draw)
-		if err != nil {
-			log.Println(err)
-			continue
-		}
-	}
-	return nil
-}
-
-func prepInsertSQLiteDrawStmt(ctx context.Context, db *sql.DB) (*sql.Stmt, error) {
-	stmt, err := db.PrepareContext(ctx, insertSQLiteDrawSQL)
-	if err != nil {
-		return nil, fmt.Errorf("%w-%s", dbutil.ErrDBPrepareStmt, err.Error())
-	}
-	return stmt, nil
-}
-
-func insertSQLiteDraw(ctx context.Context, stmt *sql.Stmt, d Draw) (sql.Result, error) {
-	result, err := stmt.ExecContext(ctx, d.DrawDate.Unix(), d.DayOfWeek, d.Ball1, d.Ball2, d.Ball3, d.Ball4, d.Ball5, d.LifeBall, d.BallSet, d.Machine, d.DrawNo)
-	if err != nil {
-		return nil, fmt.Errorf("%w-%s", dbutil.ErrDBInsertTbl, err.Error())
-	}
-	return result, nil
-}
